Create CommentFormCustomizer-class
- Move hooks from the MasterSite class
